Former-commit-id: bf1f1b9e3a446cd6899b1684f1af45a9cb9cbb33

+++ acalib/algorithms/clumpFind.py
@@ -4,8 +4,10 @@ from acalib import *
 from acalib.cupid import pycupid
 import astropy.units as u
 from astropy.nddata import *
+import numpy as np
 
 @support_nddata
 def clumpfind(data, wcs=None, mask=None, unit=None, rms=0.0):
-    ret = pycupid.clumpfind(data, dict(), rms)
-    return NDData(ret, uncertainty=None, mask=None, wcs=wcs, meta=None, unit=unit)
+	data = data.astype(np.float64)
+	ret = pycupid.clumpfind(data, dict(), rms)
+	return NDData(ret, uncertainty=None, mask=None, wcs=wcs, meta=None, unit=unit)
